requirements.txt
google-generativeai
PyMuPDF
streamlit
pytesseract
pdf2image
pillow
